bpf: Fix ICMPv6 handling in host firewall
[ upstream commit 8b62a2b6b109a52fbadf706a589b5fe4ea52dea6 ]

For ICMPv6 messages we allow, we need to bypass the host firewall. We
cannot simply pass these messages up the stack because they are not
destined to the host.

I rebased 5244b68 not long before it was merged and likely introduced
the bug at that point. I validated the rebase with an ingress host
policy instead of an egress, and missed the regression. This time I
validated with an extensive ingress+egress L3L4 host policy included at
the end of this series.

Fixes: 5244b68 ("bpf: Hande icmpv6 in host firewall")
Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Joe Stringer <joe@cilium.io>