wireguard: Add NodeIP to AllowedIPs list
This commit adds a remote NodeIPv{4,6} to the AllowedIPs to the
corresponding peer entry.

This is required, as after we have removed the Wireguard tunnel IP,
requests forwarded by NodePort BPF will have the NodeIP instead of the
tunnel IP. So the remote peer will receive such requests on "cilium_wg0"
with srcIP = NodeIP. If the NodeIP is not in the AllowedIPs list, the
request will be dropped by the Wireguard device driver.

Signed-off-by: Martynas Pumputis <m@lambda.lt>