wireguard: Encrypt pod2host traffic when NodePort BPF is on
When running Wireguard with KPR, we want to encrypt the following
NodePort BPF traffic:

    1. A request from a client to a service handled by NodePort BPF when
       a remote service endpoint is selected (after SNAT and DNAT).
    2. A reply from the service endpoint to the request above from on
       the remote node.

The 1. is marked for encryption from the NodePort BPF program (once it
detects that the service endpoint is remote). The encryption of 2. is
more complicated.

Previously, when we had a Wireguard tunnel IP, the request was SNAT-ed
to the tunnel IP. This made the reply to automatically go over the
Wireguard tunnel.  Unfortunately, after removing the Wireguard tunnel
IP, this is no longer the case. The request is SNAT-ed to a node IP, and
therefore a reply (src=podIP|dst=nodeIP) is no longer encrypted, as it
does not go over the Wireguard tunnel.

To fix this, we mark a packet sent from a pod for the encryption if we
detect that dst is REMOTE_NODE_ID. Then later on, the packet is being
routed by the stack to the Wireguard tunnel device (thanks to the mark).
Unfortunately, this introduces an asymmetry for non-NodePort host2pod
traffic, as host2pod will go unencrypted over a native device while
pod2host will go over the tunnel and will be encrypted.

Signed-off-by: Martynas Pumputis <m@lambda.lt>