fqdn: Limit the number of zombies per host
Commit X introduced a per-host limit on the number of IPs to be
associated in the DNS cache, but at the time we did not support keeping
FQDN entries alive beyond DNS TTL ("zombie entries"). These were later
added in commit Y, but at that time no such per-host limit was imposed
on these zombie entries.

Later on, commit Z adjusted the zombie garbage collection to allow
zombies to stay alive as long as any IP that shares the same FQDN is
marked as alive. Unfortunately, this lead to situations where a very
high number of DNS cache entries remain in the cache beyond the DNS TTL,
simply because one IP for the given name continues to be used.

In the case of something like Amazon S3, where DNS TTLs are known to be
low, and IP recycling high, if an app constantly made requests via
ToFQDNs policy towards names hosted by this service, this could lead to
thousands of stale FQDN mappings accumulating in the cache. For each of
these mappings, Cilium would allocate corresponding identities, and when
this is combined with a permissive pod policy, this could lead to
policymaps becoming full, and error messages in the logs like:

    msg="Failed to add PolicyMap key" ...
    error="Unable to update element for map with file descriptor 67: argument list too long"

This could also prevent new pods from being scheduled on nodes, as
Cilium would be unable to implement the full requested policy for the
new endpoints.

In order to mitigate this situation, extend the per-host limit
configuration to apply separately also to zombie entries. This allows up
to 'ToFQDNsMaxIPsPerHost' FQDN entries that are alive (ie below DNS TTL)
in addition to a further 'ToFQDNsMaxIPsPerHost' zombie entries
corresponding to connections which remain alive beyond the DNS TTL.

Signed-off-by: Joe Stringer <joe@cilium.io>