dependabot: Remove doc build system updates, Dependabot config itself
Remove Dependabot's updates for required dependencies.

The reason is that when Dependabot attempts to update
Documentation/requirements-min/requirements.txt, then it is not enough
to merge this PR. Instead, we have to synchronise with the actual list
of dependencies ("make -C Documentation update-requirements"), and to
check that everything still works as intended.

Most of the time, Dependabot's updates for documentation requirements
are just noise, because the update don't bring any useful feature, and
we don't have the cycles to sync and test the changes for each PR, this
is simply too much work for too little benefit.

Worse, some PRs might be merged without taking the time to sync and
test, for example: https://github.com/cilium/cilium/pull/22155, and in
such case further updates of the full requirements list will be broken.

So let's stick to manual updates again.

Regarding security updates, it appears that Dependabot offers them
regardless of this configuration file (by updating directly
Documentation/requirements.txt), as seen in
https://github.com/cilium/cilium/pull/23980.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>