clustermesh: introduce phantom service annotation
[ CEE commit 39cb174547bf20eb472546b157ccb858c3a556f5 ]

Marking a service (of type LoadBalancer) as "phantom" means that the
corresponding LB IP addresses become reachable from the remote clusters,
even if a service with the same namespace/name does not exist there.
This commit introduces the support for the service.isovalent.com/phantom=true
annotation to specify that a given service is "phantom".

Co-authored-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>