clustermesh: propagate phantom services info through etcd
[ CEE commit 795f93ef0c6c378ea1e381d9ae83361bbf19bab2 ]

Propagate the information concerning phantom services into etcd, so that
they can be consumed by remote clusters. In order not to modify the
current API, phantom services are identified setting IncludeExternal=false
and Shared=true (in case of global services both flags are set to true).

In addition, we advertise the LB IP addresses as frontends (instead of
the ClusterIPs as usual), as those will be used in the remote cluster to
access the service itself (since they are guaranteed to be unique across
clusters).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>