operator/dnsresolver: Add FQDN store implementation
The FQDN store is a specialized form of storage that holds the mapping
between FQDNs and the latest known list of addresses for each of them.

It offers an API to refresh the stored list of IPs. Also, it offers
other APIs to let multiple consumers subscribe to a stream of update
notifications. Upon subscription, a consumer will be notified after
each change in the store internal status, and it will then be possible
for the consumer to retrieve the store values. The notifications are
not blocking on the store side, so each consumer is allowed to handle
each notification at its own pace without affect the others.

A unit test is added to verify the correctness of the store internal
status following a stream of fqdn to IPs updates and two other tests to
verify store internal status with single and multiple subscribers.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>