operator/dnsresolver: Add a cell for resolvers manager
Add a cell for operator background resolvers manager. The manager
listens for events related to IsovalentFQDNGroup objects and for each of
them, it starts the needed resolvers (one for each FQDN in the group)
and a CIDRGroup reconciler.

In case of updates, the manager resync the resolvers, either stopping no
more useful ones or starting new ones. Finally, it recreates the
reconciler, making it listening for updates related to the new set of
resolvers.

In case of deletion, the manager removes the CIliumCIDRGroup object and
stops all the related resolvers and the reconciler.

Co-authored-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>