multinetwork: Implement /network/attachment API call
This commit adds a new cell to implement the `GET /network/attachment`
API call. When queries about the network attachments of a pod, the
current implementation looks at the
`network.v1alpha1.isovalent.com/pod-networks` annotation on the pod. It
then looks up all those networks in via the `IsovalentPodNetwork` CRD
and returns the routes for each configured network. If the annotation on
a pod is not set, we simply return the configuration for the `default`
network, which is then assumed to exist. If the pod object or network
object cannot be found, an error is returned. The API call will be
issued by the CNI plugin, which forwards that error to the container
runtime and eventually kubelet. This means that the request will be
periodically retried while the pod remains in `ContainerCreating`. This
allows the user to create the missing network or fix the faulty pod
annotation, enabling the `GET /network/attachment` call and thereby the
CNI ADD request to eventually succeed.

The API call is only available if the multinetwork cell is enabled
via `enable-multi-network` flag (false by default).

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>