workflows: enable removing label on closing issue
This workflow can close issues if they are mentioned in the PR
description in the right way. It's used in our backporting process,
which would benefit from also removing the
`enterprise-backport-pending/1.X` label from the issue we're closing.
This patch adds the ability to specify a label to be removed from the
issue prior to it being closed.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>