Make aggregator GC start an explicit call
Introduce clockwork to make it easier to manage time handling in tests

I was seeing flakes after porting the hubble-flow-aggregation logic to
cilium, and I believe it's a race, which I could reproduce in the
original github.com/isovalent/hubble-flow-aggregation.

To fix it, I eliminate the possibility of races by managing the clock
directly with a fake clock.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>