clustermesh: add phantom service tests to cilium-cli
[ CEE commit e1c4bddad8106fa386737411b3a2655da6c83e18 ]

[ backporter's notes: small conflicts in the enterprise cilium-cli
  go.mod and go.sum files, resolved accepting both changes and
  re-running go mod tidy. Additionally bumped the cilium-cli version
  from v0.15.0 to v0.15.4 due to method not available. ]

This commit leverages the hooks framework to extend the cilium CLI
connectivity tests with additional scenarios to verify that phantom
services work properly. A setup function is furthermore registered
to create the phantom service in one of the clusters, and configure
fake LoadBalancer IP addresses (selected among the ones reserved
for documentation purposes); a dummy LoadBalancer class is also set
to prevent real implementations from reconciling this service.
To prevent race conditions, the setup function waits until the service
information has been propagated to the agents running on the nodes
hosting the client pods. A finalizer is registered to remove the
service when the test completes.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>