bpf: nodeport: provide L4 ports for SNAT in LB egress path
[ upstream commit e8fa91ddbc1a54d22baa7320e60ba96b560b870e ]

For LB traffic that gets forwarded to a remote backend in non-DSR mode,
tail_nodeport_nat_egress_ipv*() calls snat_v*_nat() to perform SNAT on the
packet. Under the covers, this extracts a fresh CT tuple to look up / build
a SNAT entry.

But for LB traffic we don't require any of the ICMP handling in that code
path, and we already extract a CT tuple for building tunnel headers in XDP
mode. So we can optimize this code path, and provide a fully populated CT
tuple to the SNAT helper.

One additional benefit is that we fix handling for fragmented IPv4 packets,
as lb4_extract_tuple() knows how to extract their L4 ports while
snat_v4_nat() doesn't.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>
Signed-off-by: Jussi Maki <jussi@isovalent.com>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>