clustermesh: handle "delete cluster while Cilium is stopping" case
[ CEE commit 064bf4488c8920c3866256a077694e8712012a56 ]

Cleanup stale per-cluster maps that were bound to possible clusters
deleted while Cilium agent is stopping. When we perform the cleanup, we
first wait for the initial sync of the clusters and wait for at least
all of the remote clusters on the configuration file finish reserving
the required ClusterIDs. After that, we perform cleanup with the
usedClusterIDsMutex held. So that we can avoid the scenario like,

1. New cluster added while we are cleaning up.
2. A controller of the new cluster reserves the ID and create
   per-cluster maps.
3. Right after that, cleanup for the acquired ID happens and we break
   the connection of the new cluster.

Co-authored-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>