maps: implement a wrapper around per-cluster maps
[ CEE commit 3ce8f07be44acf170a03039197c8942cb309b751 ]

Introduce a new maps.PerCluster interface which abstracts the logic to
manage the per-cluster CT/NAT maps used for overlapping PodCIDR support.

The constructor function takes care of initializing the per-cluster maps
when users enable cluster-aware-addressing. When users disable it, it
always tries to cleanup the per-cluster maps, because users may enable it
first and disable it after that. We also handle the cases that users
first enable IPv4/IPv6 and change it afterwards.

Finally, the constructor provides the appropriate implementation for
the PerCluster interface, depending on whether cluster-aware-addressing
is enabled.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>