k8s: Add IsovalentPodNetwork CRD and scaffolding
This generates the IsovalentPodNetwork and IsovalentPodNetworkList
types, to specify a network for Cilium-native multi-homing.

The initial version of the CRD is intentionally kept simple: It only
contains the IPAM pool (to be used for IP allocation on that network)
and a list of routes that define the network from a routing persective.
The list of routes is used to steer pod traffic to that network in a
multi-homed pod.

Co-authored-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>