operator: propagate CiliumClusterConfig when in kvstore mode
[ upstream commit 45174b3c97779a4288b59f6650d4d4b7d6380418 ]

5e5a26e5da4d ("clustermesh: Implement a basic connect-time validation")
introduced the CiliumClusterConfig, which is stored in the kvstore by
the clustermesh-apiserver at startup time to propagate information about
the current cluster (e.g., its cluster ID). The same operation, though,
was not performed when cilium is configured in kvstore mode (the
clustermesh-apiserver is not enabled in this case). Let's fix this
deviation implementing the same logic also in the cilium operator.
No capabilities are set in this case, because sync canaries are not
supported (and not needed given that the kvstore is expected to leverage
persistent storage in this case).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>