bgpv1: Add generic Path type
[ upstream commit 6195047ba1759856c06b419d6ecd2e92e6f76d5d ]

Add a new generic Path type that which is an analogue of the GoBGP's
Path object, but only contains a minimal fields required for Cilium's
usecases. It consists of NLRI, Path Attributes and some status fields.

We are relying on the GoBGP's pkg/packet/bgp package which is a generic
BGP library (say net.IP package, but for BGP). We decided to reuse this
library as a part of our router-independent intermediate representation
because it's too painful for us to reinvent the BGP library by
ourselves.

This type will be used for implementing various BGP CPlane
functionalities like retrieving the route.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>
Co-authored-by: Rastislav Szabo <rastislav.szabo@isovalent.com>