kvstore: remove synced variable from RestartableWatchStore
[ upstream commit fd9b1e60857aac712c26f4f50ba685242b2dd266 ]

Currently, the synced variable is used in the RestartableWatchStore
to ensure that callbacks are executed only once upon the initial
synchronization. Given that a subsequent commit will start exposing
synchronization information (with a different meaning), let's get rid
of it and rework the callbacks execution to not depend on it.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>