ci: skip cosign / sbom in case of building images during cache rebuild
[ upstream commit 56960c74488939930e819a91d0b1d75718f5e444 ]

Docker cache re-build is triggered via event `workflow_run`. In this case,
the container image isn't puhed to a docker image, therefore
`outputs.digest` isn't set. This results in cosign and sbom steps
to fail because they depend on this.

This commit fixes this by only executing these steps in case of a github
event of type `push`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>