ci: set nullglob when trying to copy junit testfiles
Trying to copy the junit testfiles via wildcard results in the following
error if no testfiles exist.

`cp: cannot stat '*.xml': No such file or directory`

This is due to the fact, that bash passes the wildcard as a literal if
it can't find any file matching the pattern.

Therefore, this commit sets the option `shopt -s nullglob` to prevent
this from happening.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>