endpoint: Allow multiple endpoints per pod
[ upstream commit ced01f656f03149e58065ad1dcbefbcb833009cd ]

This commit removes the restriction that there is a 1:1 mapping between
a pod and an endpoint. To achieve this, we remove the `LookupPodName`
helper in the agent and replace it with a `LookupCEPName` (for
correlating with a CEP) and `GetEndpointsByPodName` (for correlating
with a pod) helpers instead.

Lookups on the pod name via API continue to be supported for backwards
compatibility, if the endpoint has not opted out of legacy identifiers.
The API also allows lookup by `cep-name` with this change, but similar
to other mostly internal identifiers (such as the `cilium-global`
prefix) we do not announce this ID in the `EndpointIdentifiers` model.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>