api: Add `cni-attachment-id` as endpoint identifier
[ upstream commit 23da0ad72dba98943f8618d5ec1735f7ce77fb0f ]

This commit adds a new endpoint identifier called `cni-attachment-id`.
It it intended to replace the container ID and container name as a
unique identifier, because the CNI spec allows the runtime to invoke a
CNI ADD/DEL for the same container ID twice. Therefore, instead of the
container ID, we now recommend using the CNI attachment ID, which
consists of the `(containerID, ifName)` pair.

The container ID and container name IDs will continue to work for
primary attachments, but are now deprecated.

This commit only changes the API definition. The new ID will be hooked
up in a subsequent commit.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>