cni: Follow CNI spec by using `(containerID, ifName)` as endpoint ID
[ upstream commit 259f8f8eb343f87542f4b4a6649a5d1d43e09f71 ]

The CNI spec states [1] that a attachment is uniquely identified by the
tuple consisting of both the container runtime container ID and the
container interface name. Since a Cilium endpoint corresponds to a
network attachment, the container ID alone is not sufficient to uniquely
identify an endpoint.

Therefore, this commit deprecates the concept of the container ID as a
unique identifier for an endpoint, and instead introduces the "CNI
attachment ID", which consists of both the containerID and container
interface name. This is guaranteed to be unique according to the CNI
spec.

The container id (and container name) continue to be supported in the
Cilium API when an endpoint did not opt out of those legacy identifiers,
in which case it is also implicitly assumed that there is only one
endpoint per container.

[1] https://www.cni.dev/docs/spec/#section-3-execution-of-network-configurations

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>