api: Deprecate `pod-name` as endpoint id
[ upstream commit 6f86237c369e84d8113d4f5817a06e5fcfda8448 ]

This commit deprecates the `pod-name` lookup in the Cilium API and
replaces it with a `cep-name` lookup instead. As with the previous
change in this commit series, the `pod-name` may no longer uniquely
identify an endpoint if a pod has multiple network attachments.

`pod-name` will continue to work as an endpoint id for the endpoint of
the primary network interface.

This commit only changes the API definition. The new ID will be hooked
up in a subsequent commit.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>