ci: fix Azure cluster names sometimes being too long
[ upstream commit 0297c6c6ce4d57a596a74e092256442b1f222c80 ]

Azure does not allow having multiple clusters with the same name in the
same subscription even if they are hosted on different locations.

In order to avoid name conflicts, we previously added the location name
to the cluster name, however in some cases this leads to cluster names
exceeding the maximum length.

As a quick fix, we replace the location name with a simple index.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>
Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>