ipcache: Stop populating node IDs in ipcache
[ upstream commit 4856e3829193351712c1fc4958639bc34bbdf724 ]

In previous commits, we removed all consumers (IPsec and Mutual Auth)
of the ipcache's node_id field. This commit removes population of those
node IDs in the ipcache.

**Problem**

Needing to populate the ipcache with node IDs causes all sorts of
issues. Node IDs are allocated for and mapped with specific node IP
addresses. In practice, it means node IDs are tied to the lifecycle of
both node and pod objects:
1. We need to allocate node IDs when receiving node objects with new
   node IP addresses (either new node objects or existing node objects
   with changing IP addresses).
2. We need to allocate node IDs when receiving a node IP address via a
   pod (hostIP of the pod).

Node objects have all node IP addresses. So why are we not relying
solely on 1? We also need 2 because pod adds and pod updates (with new
hostIP) can arrive before the related node add or node update.
Basically, kube-apiserver doesn't provide any guarantee of ordering
between node and pod events. Since we need to populate the ipcache with
a node ID for each remote pod, we need to allocate a node ID when we
receive the pod object and not later, when we receive the node object.

Unfortunately, tying the node IDs to two object lifecycles in this way
causes a lot of complications. For example, we can't simply deallocate
node IDs on pod deletions because other pods may still be using the node
ID. We also can't deallocate simply on node deletions because pods may
still be using it. The typical solution to such issues is to maintain a
reference count, but that leads to its own set of issues in this case
[1].

**Considered Solution**

An alternative to all this is to force events to always be managed in
the same order. For example, NodeAdd -> PodAdd -> PodDelete ->
NodeDelete. The agent unfortunately doesn't easily lend itself to that.
Without a major refactoring, we would have to populate the ipcache as
usual, but skip node IDs. Then, on NodeAdd, we would allocate a node ID
and go through the ipcache again to populate the node IDs we skipped.
In the datapath, we would have an explicit drop for such half-populated
ipcache entries.

To implement this without needing to walk through the entire ipcache for
each NodeAdd or NodeUpdate, we need to keep track of the specific
entries that need to be populated. We need the same sort of mechanism
for deletions. This solution quickly turns into something quite complex
to implement with good performance.

**This Solution**

It should be quite clear at this point that this problem's complexity
stems from having node IDs tied to both the pods and nodes' lifecycles.

To untie node IDs from pod lifecycles, and thus, stop populating the
ipcache with node IDs, we will need the datapath to retrieve them from
some other place. Fortunately, we already have a BPF map that maps node
IP addresses to node IDs. It is currently only used as a restoration
mechanism for agent restarts, but there's no reason we can't use it for
more.

Thus, with the node IP address retrieved from the ipcache
(ep->tunnel_endpoint), we can lookup the node ID in the BPF map.

Of course, since pod objects can arrive before their corresponding node
object, the node ID may not always be ready in the BPF map when we need
it. When that happens, we will drop the packet with a specific drop code
(DROP_NO_NODE_ID). These drops simply mean we tried to send traffic to a
node for which we didn't receive the node object yet (hence the
encryption configuration isn't setup yet).

These datapath changes have been implemented in the previous commit.
This commit removes all population of node IDs in the ipcache.

We can also remove node IDs from the tunnel map. It matters less since
the tunnel map is tied to node objects, but let's keep the datapath
consistent between the two maps (ipcache and tunnel map).

1 - https://github.com/cilium/cilium/pull/26725#discussion_r1266785189
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Tam Mach <tam.mach@cilium.io>
Signed-off-by: Andrew Sauber <2046750+asauber@users.noreply.github.com>