bpf/tests: Rename IPsec unit tests
[ upstream commit 426c42495c91be0f4f5c2bdbefdc6ff9932fcc33 ]

This will be useful to enforce ordering between different tests, with
the first one preparing part of the maps for subsequent tests. Numbering
starts at 2 because the next commit will introduce a test 1.

"ipsec_" is removed from the test names to shorten it a bit (the IPsec
context should be clear from the test filename).

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Tam Mach <tam.mach@cilium.io>
Signed-off-by: Andrew Sauber <2046750+asauber@users.noreply.github.com>