node: Fix leak of node IP <> ID mappings
[ upstream commit 4d46fa9c9def58396f252a1cfdefe0a820e233fd ]

This commit fixes a leak of node IP to node ID mappings when the set of
IP addresses of a node is updated. We have logic to allocate and map
node IDs on node creation, logic to unmap and deallocate node IDs on
node deletions, but we didn't fully handle the case of node updates.

On node updates, we would map new IP addresses to the already-allocated
node ID. We forgot to unmap the old IP addresses. This commit fixes that
oversight.

This leak means that some IP addresses that were previously assigned to
the node would remain mapped to the node ID. Those IP addresses could
then be assigned to a new node. The new node would then receive the same
node ID as the old node (because we first check if any of the node IPs
already have an ID). This could lead to encrypted traffic being sent to
the wrong node.

Before removing the node IDs from the ipcache (see previous commit),
this leak could have another consequence. If after the node ID had been
reassigned to the new node, the old node was deleted, the node ID would
be deallocated, but it would remain in the ipcache. The XFRM policies
and states of the node would also be identified based on the node ID and
removed. Thus, both the old and new nodes' XFRM configs would be
removed. As a consequence, outgoing packets for those nodes would be
dropped with XfrmOutPolBlock (the default drop-all rule).

Fixes: af88b42bd ("datapath: Introduce node IDs")
Signed-off-by: John Fastabend <john.fastabend@gmail.com>
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Tam Mach <tam.mach@cilium.io>
Signed-off-by: Andrew Sauber <2046750+asauber@users.noreply.github.com>