nodemap: Extend fake node map for use in tests
[ upstream commit 52b0430c9ef7701d2a57cad91455cea382c7010f ]

The fake node ID map currently only defines the minimum set of functions
needed for the interface. This commit extends it to simulate the actual
behavior of the BPF map with an in-memory Golang map.

This will allow us to use check that the map content matches our
expectations for various events of the node object lifecycle.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Tam Mach <tam.mach@cilium.io>
Signed-off-by: Andrew Sauber <2046750+asauber@users.noreply.github.com>