node: Integration test for node ID lifecycle
[ upstream commit fc8da7266ab03b6813cb7469e2fd217187308acc ]

The node ID lifecycle is tied to the node lifecycle. This commit adds an
integration test to cover this lifecycle. It covers the creation and
deletion of nodes, as well as the update of a node's IP addresses. It
also checks that two nodes don't end up with the same node ID and that
each node has a single node ID for all its IP addresses.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Tam Mach <tam.mach@cilium.io>
Signed-off-by: Andrew Sauber <2046750+asauber@users.noreply.github.com>