api: Add flag to disable legacy identifiers
[ upstream commit fec254f8af030b486fe1297480c01a202598c4c3 ]

This will allow endpoints to opt-in to new endpoint identifier system,
which requires endpoints to be looked up either using the CNI attachment
ID or the CEP name (both of which set the container interface name).
This will allow multiple endpoints within the same pod or container.

If an endpoint creation request does not explicitly opt-in, we will
continue to support addressing it by the container ID or pod name alone
for backwards-compatibility reasons.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>