maps: refactor per-cluster CT maps manager
[ upstream commit e6a7df47243f03144124c4f49537bd6f7d18afc1 ]

Let's simplify the per-cluster CT maps management logic, removing the
dependency on the global variable (which does not play well with the hive
framework) and splitting the creation/removal tasks from the retrieval
of inner maps.

While being at it, let's also slightly improve the performance around
these operations by avoiding to attempt opening the inner maps when
not strictly necessary, as well as keeping the list of known Cluster IDs.

Additionally, let's also move the fake implementation for testing purposes
to a separate package, for better separation.

Finally, let's validate the ClusterID in the `bpf ct list cluster <ID>`
CLI command, to properly return an error in case it is invalid.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>