maps: do not depend on global variable to initialize CT maps
[ upstream commit 69b2d3a7de1f4cb2aa12b5cfb5313f11ec0bf67f ]

Currently, the creation of a new CT map depends on the `mapInfo` global
variable. Yet, this is suboptimal, as it always requires that variable to
be initialized beforehand, which might be tricky in combination with the
Hive framework. Additionally, all the necessary information is either
known at compile time or depends on user configurations only. Hence,
let's update that logic to remove the dependency on the global variable.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>