multinetwork: Detect and announce secondary node IPs
This is the first of two commits which adds support for a multi-network
aware auto-direct-node-route feature to Cilium-native multi-homing.
Direct node routes installs a route in the form of `$podCIDR via
$nodeIP` for each remote node in the cluster, thereby allowing pod
traffic to be routed to the connect node on a L2-connected network.

In multi-homing mode, we now need to assume that a node will have
multiple node IP addresses, i.e. at least one per network. Therefore,
this commit adds a new controller which periodically lists the host's
network interfaces and extracts node IP addresses based on the route
listed in the `IsovalentPodNetwork`. If the interface IP matches one of
the network's routes, we then assume it is the canonical IP address of
the node in that network. Future extensions and configurations to this
mechanism are possible, but for now we just support this rather simple
mechanism.

These secondary node IPs are then passed to the `NodeDiscovery` package,
which is responsible for announcing them to other nodes in the kvstore
and the CiliumNode CRD. This also has the nice side effect that other
cilium-agents will now add those secondary node IPs into the IPCache
with the `remote-node` identity.

In order for the node kvstore schema to remain compatible to the
upstream one, we currently store the network name in the address type
field. This is a temporary solution, the goal is to eventually upstream
parts of this commit and change the node schema to store the network
name.

When installing the direct node routes (see first paragraph), we will
then pick the `$nodeIP` based on the network the `$podCIDR` belongs to.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>