cni: Add source based routes for cross-network traffic
If there is cross-network traffic, the network routes can cause the
reply packet to leave on a different interface than the one it was
received on. This causes Cilium to drop the reply packet because the
packet on the egress interface does not match the egress interface's
source IP (Cilium performs endpoint source IP verfication in the
datapath).

To solve this, we install a source IP based IP rule that redirects
packet with a pre-determined source IP over the interface where that IP
is attached to.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>