maps: provide CT/NAT maps GC logic through Hive
[ upstream commit e38e0ba0beebc3823e3a33fd2562feb8a8f63247 ]

Refactor the CT/NAT maps GC logic into an object, which is provided
through Hive. While it is not yet possible to start the actual logic
through Hive (as it depends on the list of regenerated endpoints,
which is still constructed in the daemon constructor itself, and the
CT/NAT maps have not yet been converted into cells), this makes
adding additional dependencies easier, without having to increase
the number of parameters of the Enable function.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>