pkg/node: expose 'newTestLocalNodeStore' for testing
[ upstream commit a8d80e25ca2d26affe4f7f7a8350bb537ddbe388 ]

This commit makes the previously private 'newTestLocalNodeStore'
constructor public and also updates the constructor to receive a mock
LocalNode structure.

The returned 'LocalNodeStore' will then "emit" the mock LocalNode when
invoked.

This is useful for unit tests which include the 'LocalNodeStore'
structure as a depedency, avoiding the need to create a Hive for the
unit test.

These changes will be used in subsequent commits.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>