cilium: trigger validation of all nodes when key updated
[ upstream commit bba3dfcec7b5626fde845fab4df094275d30c371 ]

From: John Fastabend <john.fastabend@gmail.com>

On key update we only update the policy for our local node. But with
latest round of changes we need to update for all nodes in the node
cache. With out this we would rely on the validation interval timer
to sync the nodes policy before we remove the old policy. This
may or may not happen depending on how large the cluster is.

Further I've seen us miss it even on relatively small clusters say
around 30 nodes so seems its not entirely reliable to count on.

Rather than rely on some external to ipsec timer to fire and sync
the policies and to do it hopefully in our time window lets just
force the nodeUpdate() call on all nodes in the cache when we
get the key rotate event.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>
Signed-off-by: Zhichuan Liang <gray.liang@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>