gateway-api: Filter routes based on Section Name and port
[ upstream commit 5e1dff0130a6788d680fc37b9df5d370c1ae0705 ]

This is to make sure that we are having a correct attached routes in gateway
status. The translation logic is already implemented such filters, but it
was overlooked in Gateway status.

For the below spec, the attached routes for both listeners was showing as 1,
the expected behavior should be 0 for http-non-attached listener, and 1 for
http listener. Similar scenario should be applied for port number.

```yaml

Signed-off-by: Tobias Klauser <tobias@cilium.io>
---
apiVersion: gateway.networking.k8s.io/v1beta1
kind: Gateway
metadata:
  name: sample
spec:
  gatewayClassName: cilium
  listeners:
    - name: http-unattached
      port: 8080
      protocol: HTTP
      allowedRoutes:
        kinds:
          - kind: HTTPRoute
        namespaces:
          from: All
    - name: http
      port: 80
      protocol: HTTP
      allowedRoutes:
        kinds:
          - kind: HTTPRoute
        namespaces:
          from: All
---
apiVersion: gateway.networking.k8s.io/v1beta1
kind: HTTPRoute
metadata:
  name: http-route-4
spec:
  parentRefs:
    - kind: Gateway
      name: sample
      sectionName: http
  rules:
    - backendRefs:
        - name: infra-backend-v1
          port: 8080
```

Signed-off-by: Tam Mach <tam.mach@cilium.io>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>