docs: Replace mentions to KPR strict mode for per-node-config docs
[ upstream commit d97cd7a82a774bf66b11840668de2eb4098e3830 ]

The "strict" and "partial" mode for kube-proxy replacement (KPR) were
deprecated in Cilium 1.14, in favour of the feature being enabled (the
flag is set to "true" and all related options are turned on) or disabled
(leaving users free to pick the options they want).

Let's update the documentation for per-node configuration accordingly
and use "true" instead of "strict".

Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>