resource: Fix double upserts on subscribe and retrying of delete events
[ upstream commit a7f1eeffc1c63d0557207f780ad453fa177613b7 ]

[ Backporter's notes: adapted the patch to work without
  7959c43965f2c193e55d211519c0b5d269d67ac3 and
  9cdcddc67061c59df7b517ef52a1d5a8f0b700c1. The adaptions
  should not pose difficulties in backporting them. ]

Fix double upserts that were caused by store being manipulated without
synchronization with the subscriber queues by processing the deltas
under the resource read-lock and doing the initial key listing for
new subscriber with a write-lock. This way we cannot accidentally
see a key in the store and process it just before the key is queued.

As shown by test case in previous commit, the delete events are retried
with an old incorrect version of the object causing a recreated object
to be deleted.

Fix the deletion retrying by always queueing upserts and deletes by key
and keeping the last known state of objects emitted to the subscriber.
Only emit a delete event if the subscriber has seen its creation and
only use a version of the object that the subscriber has observed.

Fixes: 4101e2c768 ("k8s: Add resource package")
Signed-off-by: Jussi Maki <jussi@isovalent.com>