egressgw: add test for conflicting IP rules
When the DestCIDR of a policy changes, we first add the new IP rules before
deleting the old ones. But if the new DestCIDR is 0.0.0.0, attempting to
install the matching IP rules results in an -EEXIST error.

To deal with this, addMissingIpRulesAndRoutes() returns a `shouldRetry`
value. If it is set, we retry the installation after
removeUnusedIpRulesAndRoutes() has deleted the unused rules.

Add a test for this scenario.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>