datapath: allow injecting defines into NodeConfig via a map
[ upstream commit 9924a287e1cd79029620337a928126cb85eecaa6 ]

The existing functionality to inject defines via a function is
necessary for cases where the defines are not know at construction
time. If they are known, it's nicer to be able to specify a plain
map. This gives quicker feedback if duplicate defines are attempted.
It also allows leaving the map nil if a feature is disabled, where
in the existing implementation you'd need a dummy function.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>