.github/workflows: add issue closer workflow
GitHub only auto-closes issues via PRs with 'Fixes: #xyz' in the
description if the PR targets the default branch of the repository.
That's incompatible with the stable branch model, so let's use our
reusable close-fixed-issues.yaml workflow on default.

The Github action to trigger this needs to live in each enterprise
branch, since the pull_request trigger only works in the base/target
branch of the PR.

Like on the other v1.x-ce branches, add a workflow that closes issues
fixed by v1.14-ce backport PRs.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>