bpf: lxc: defer CT_INGRESS entry creation for loopback connections
[ upstream commit dbf4351c4eff1857cf374886804eb876e970b992 ]

[ backporter's notes: bring back ipv4_ct_tuple_swap_addrs() to avoid churn
  in the tests. ]

Currently the CT_INGRESS entry for a loopback connection is already created
when the first request leaves the client, as from-container creates the
CT_EGRESS entry (see the loopback handling in ct_create4() for details).

This is unusual - we would normally just create the CT_INGRESS entry as
the first packet passes through to-container into the backend. But for
loopback connections it is needed, so that
1.) to-container finds a CT entry with .loopback set, and thus skips
    network policy enforcement even for the first packet, and
2.) the CT entry has its rev_nat_index field populated, and thus can
    RevNAT replies in from-container.

This approach conflicts with the fact that loopback replies skip the
client's to-container path (to avoid network policy enforcement).

Once the loopback connection is closed, the backend's from-container path
observes the FIN / RST, and __ct_lookup4() updates the CT_INGRESS entry's
lifetime to CT_CLOSE_TIMEOUT. But the client's to-container path will not
observe the FIN / RST, and thus the CT_EGRESS entry's lifetime remains as
CT_CONNECTION_LIFETIME_TCP. Therefore the CT_INGRESS entry will expire
earlier, and potentially gets garbage-collected while the CT_EGRESS entry
stays in place.

If the same loopback connection is subsequently re-opened, the client's
from-container path finds the CT_EGRESS entry and thus will *not* call
ct_create4(). Consequently the CT_INGRESS entry is not created, and the
backend will not apply the loopback-specific handling described above.
Inbound requests are potentially dropped (due to network policy), and/or
replies are not RevNATed.

Fix this up by letting the backend path create its CT_INGRESS entry as
usual. It just needs a bit of detection code in its CT_NEW handling to
understand that the first packet belongs to a .loopback connection, and
populate its own CT_INGRESS entry accordingly.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>
Signed-off-by: David Bimmler <david.bimmler@isovalent.com>