clustermesh: rework wait for the initial services synchronization
[ upstream commit 39beeb4ea339c36be058090f905646e5d8a51b11 ]

[ Backporter's notes: small conflicts due to the store factory and
  the extra ipcache options injection logic not being present in v1.14,
  and the different name of the clustermesh/common package (was
  clustermesh/internal). ]

[ backporter's notes: small conflicts while cherry-picking into
  v1.14-ce because the logic mentioned above was backported to that
  branch. Additionally adapted the usage of ClustersSynced() in the
  enterprise code. ]

Generalize the logic currently used to wait until the initial list of
shared services has been received from all remote clusters, and
synchronized with the BPF datapath, to prepare for the subsequent
introduction of similar methods for other resources.

While being at it, let's also rename the function to clarify its scope
(as it only waits for the synchronization of services), and fix the
possibility of a deadlock in case a remote cluster is disconnected
before that the synchronization has completed.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>