clustermesh: allow to wait for ipcache and identities synchronization
[ upstream commit 6c4e693d1805a6a0e172e3011801d1e751590199 ]

[ Backporter's note: small conflicts due to the store factory and the
  extra ipcache options injection logic not being present in v1.14. ]

[ backporter's notes: small conflicts while cherry-picking into
  v1.14-ce because the logic mentioned above was backported to
  that branch. ]

Add a new IPIdentitiesSynced method to the clustermesh object to allow
waiting until the initial list of ipcache entries and identities has
been received from all known remote clusters. This mimics the
ServicesSynced() method, and enables to postpone disruptive operations
to prevent dropping valid long-running connections on agent restart.

Differently from the services case, we don't need to separately
track the propagation of the single entries to the datapath, as
performed synchronously.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>