cgroups: Fix race to load cgroup.hostRoot option
[ upstream commit 603235d0b8152910800c2ef3d5a3df10cc011105 ]

While loading, the package 'pkg/cgroups/manager' sets the global variable 'cgroupRoot' from the 'pkg/cgroups' package.
This occurs before the configuration is fully loaded and the correct path is set in the 'pkg/cgroups' package.

This variable is later used by the 'validateCgroupPath' function.
As a result, the 'validateCgroupPath' function always works with the default value instead of the user-defined one.

Signed-off-by: Andrei Kvapil <kvapss@gmail.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>