multinetwork: Add multi-network aware direct node routes
This is the second of two commits which adds support for a multi-network
aware auto-direct-node-route feature to Cilium-native multi-homing.

This part adds a new CiliumNode chain subscriber which watches all
remote CiliumNodes in the cluster, extracts the remote node's multi-pool
pod CIDRs, and then installs routes in the form of `$podCIDR via
$nodeIP` for each node.

Because the CiliumNode CRD only contains the IPAM pool name of a
podCIDR, the code performs a reverse lookup on the `IsovalentPodNetwork`
list to find the corresponding pod network. This works because we
require an IPAM pool to be used by at most one network. Once the network
name of a pod CIDR is known, we can then extract the node's IP in that
network in its `Addresses` section (c.f. previous commit) and install
the routes. To ensure routes are not accidentally deleted by the user,
we also run a reconciliation controller that recomputes all routes for
each node and reinstalls them if necessary.

Ideally, we would integrate this logic into the existing
`auto-direct-node-routes` code in upstream. However, because making the
that code multi-network aware is rather intrusive and would require
kvstore schema changes, we instead let the logic live in the
multinetwork cell for now. One shortcoming of this (besides code
duplication) is that the multinetwork version only subscribes to
CiliumNode updates, which means clustermesh is not supported. However,
since multi-pool IPAM at the moment also does not support kvstore (c.f.
cilium/cilium#26621), and multi-pool IPAM is a prerequisite of
multi-network, we do not consider this a blocker.

Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>