bgpv1: Introduce BGPCPResourceStore for BGP CP reconcilers usage
[ upstream commit 58a6c71da3ae551f2d0d4c34d770e5011632ff7f ]

Introduces a new BGPCPResourceStore that provides and easy to use interface
to a resource.Store for k8s objects that needs to be processed by the
BGP Control Plane reconcilers. It automatically signals the BGP Control Plane
whenever an event happens on the resource, so that individual reconcilers do
not need to do that themselves. Also provides a respective mock type that can
be used in unit tests.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>