sidmanager: Extend SIDManager with BehaviorType
[ upstream commit ce02070b9fdde57b491fdb66790b8a5643a45c0c ]

Implement the notion of BehaviorType to SIDAllocator and SIDManager.
SIDAllocator now takes BehaviorType on construction and checks whether
the behavior provided by Allocate/AllocateNext are matched to the
BehaviorType bounded to itself. Also, Allocate/AllocateNext are modified
to return SIDInfo instead SID as we want to return BehaviorType
information as well.

SIDManager's logic is also extended to take BehaviorType into account.
For exmaple, now locator allocations in the IsovalentSRv6SIDManager can
have behaviorType fields. When it is modified, SIDManager performs
reallocation of SIDs.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>