endpoint: Remove GetLabelsLocked
[ upstream commit 08903e00e0fdbc383e8d041da08e4db4409b2796 ]

Endpoint.GetLabelsLocked existed on the premise that selector cache could
not be locked while endpoint has been locked, due to selector cache
locking ip cache in some code paths. This does not seem to be correct, as
ip cache calls in to selector cache, not the other way around.

With this the Endpoint.GetLabelsLocked can be removed, and
SelectorCache.GetLabelsLocked can be used instead also when calling from
the Endpoint locked state. This is also in line with the commend on
policy DistillPolicy that states that "PolicyOwner (aka Endpoint) is also
locked during this call", and then within takes the Selector Cache read
lock.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>