k8s: Remove check for permissive CCNPs
[ upstream commit e46ade2408553ba8bcb3cde547f3b115193994c6 ]

This check was designed to inform users of CCNP that the behaviour
changed around the timeframe of v1.9. However, the actual check is
complaining about policy that is valid. At this point there is no need
to warn users during preflight checks that they are using empty endpoint
selectors in CCNPs, because this can validly be used to select all Pods
within the cluster. Remove the check.

Signed-off-by: sh2 <shawnhxh@outlook.com>
Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>