helm: fix envoy daemonset loglevel with multiple verbose debug groups
[ upstream commit 9d0ffdff268d86036e42c3e1aae1d3c96004887b ]

Currently, the Envoy DaemonSet loglevel is not set correctly if multiple
verbose debug groups are passed as helm values. Comma is used as group
separator instead of a space. (Got confused because Go Viper flags are
separated by a comma.)

This commit fixes this by changing the separator.

Backporting conflicts:
* minor conflict as we don't check the.Values.debug.enabled condition in
  main

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>