bgpv1: IOS-XR compatible SRv6 VPNv4 advertisement
Advertise VPNv4 route with proper SID structure, SID behavior and
transposition encoding when the SID is allocated from SIDManager. When
the VRF doesn't allocate SID from SIDManager, the behavior remains the
same as before (SID structure with [128, 0, 0, 0] and End.DT4 behavior).

This aims to be interoperable with Cisco IOS-XR's SRv6 implementation.
The interoperability is verified with XRd controlplane (containerlized
IOS-XR without dataplane) 7.7.1.26I. We confirmed that the route
advertised by us are accepted by IOS-XR and the route is installed to
the FIB.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>