ipsec: Do not use AllocCIDR with subnet encryption
[ upstream commit a4c43f358eee14c717f88b26386d458a6ddfd597 ]

IPSec relies on the "pod subnet" CIDRs for encryption in IPAM modes
(e.g. ENI, Azure) which do not have a pod CIDR (aka. alloc CIDR).

Therefore, we should not use the alloc CIDR in such modes. This commit
moves the setup of routes based on the alloc CIDR into the conditions
which are executed if we do are not in pod subnet based encryption.

In addition, we add some temporary code to remove the old bogus route
based for the local node in subnet encryption mode.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: David Bimmler <david.bimmler@isovalent.com>