multinetwork: Skip installing routes without a gateway
Direct node routes consist of two parts: The pod CIDR for which we want
to route traffic, and the node IP where we want to route the traffic to.

Because these to components are coming from different sources (the pod
CIDR is assigned by the operator, the node IP is announced by the node
itself), it can happen that we know a node's pod CIDR, but not yet the
node's IP. If we then try to install such a route with a gateway (read:
node IP), it will fail with the following error message:

```
"Failed to install route for node, traffic to that node will be disrupted"
error="no such device"
nodeName=cilium-worker2
route="{Ifindex: 0 Dst: 192.168.0.96/27 Src: <nil> Gw: <nil> Flags: [] Table: 0 Realm: 0}"
subsys=multi-network
```

Eventually, we would learn about the node IP and correctly install the
route, but the warning was emitted regardless. This commit silences that
unnecessary warning by skipping the installation of any routes which do
not contain a gateway.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>