bgpv1: Provide enterprise-only reconcilers within EnterpriseControlPlane
If we provide any enterprise-only constructor with BGP Control Plane
Cell, it introduces an unnecessary enterprise dependency to the parent
ControlPlane Cell because such constructors may depend on the
enterprise-only Cells which may only provided within
EnterpriseControlPlane.

In the subsequent commits, we will make such a dependency. Namely, we
introduce ExportSRv6LocatorReconciler and it depends on SIDManager which
is only provided within EnterpriseControlPlane Cell. This has no problem
when we run with enterprise daemon (which includes both of ControlPlane
and EnterpriseControlPlane), but it breaks the existing tests under
test/controlplane which internally create their own Hive with
ControlPlane, but not with EnterpriseControlPlane.

To solve this problem, introduce a new Cell "enterprise-bgp-cp" and
provide enterprise-only reconcilers through it. The "enterprise-bgp-cp"
Cell is only provided to EnterpriseControlPlane Cell. Thus, controlplane
tests never include enterprise-only reconcilers into their dependency
tree. The Cell's group logic still works as usual when we run with
enterprise daemon.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>