ci: add enterprise multi-network workflow
This tests the Cilium-native multi-network (or multi-homing) feature in
enterprise by running connectivity tests, verifying that CEPs got
created for multi-homed pods and that these have their routes set up for
the secondary network.

Because a secondary docker network is needed for the tests, but that
cannot be created when using the helm/kind-action action, a LVH VM is
used in which the kind cluster and the secondary docker network are
created.

Currently, the workflow is triggered on every pull request/push. A
future change will convert this to an ariane-based workflow. It seems
ariane does not work for workflows that are only added by the PR itself.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>