helm: Enable enterprise-cilium-operator clusterrole
This commit moves all enterprise changes from the `cilium-operator`
clusterrole to the `enterprise-cilium-operator` cluster role, which is
maintained in it's own file for mindfulness. After this commit, the
`clusterrole.yaml` file should be identical to the one in v1.14 OSS.

In addition, the cluster role is bound to the `cilium-operator` service
account, so the role can actually be used by the operator.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>