daemon: Add option conntrackGCMaxInterval
[ upstream commit 3ea5e8c5e9c3e77b23eea30ed08200ade03e2d98 ]

[ backporter's notes:
  * dropped changes in install/kubernetes/Makefile as the given line
    was not present in v1.14;
  * fixed trivial conflict in spelling_wordlist.txt;
  * re-generated the cilium/values.yaml and cmdref/cilium-agent.md
    files from scratch.
]

When ToFQDN policies are in use (CIDR) identities are created for each IP
in a DNS response that matches a ToFQDN policy. These identities are garbage
collected when

  1) all endpoints with ToFQDN policies are removed
  2) the maximum number of IPs per host is reached (tofqdn-max-ips-per-host)
  3) when the identity has been unused and not refreshed by conntrack map GC

Problems arise when he conntrack GC interval becomes very long. If it's backed
by a LRU BPF map, the maximum is set to 12 hours (defaults.ConntrackGCMaxLRUInterval),
meaning it will take 12 hours before unused identities are marked dead and collected
(unless tofqdn-max-ips-per-host limit is reached for the FQDN entry).

To allow user to have more control over this add the conntrackGCMaxInterval option
that will allow limiting the maximum interval to something less than the 12 hours.

Signed-off-by: Jussi Maki <jussi@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>