workflows/ipsec: Add missing --flush-ct for key rotation
[ upstream commit 751c17cbdc3063ac3fac4297d33d80c1b9d49829 ]

Now that we cover the key rotations in the IPsec e2e tests, we are
running the connectivity test suite twice. That means we can run in the
usual bug where an existing CT entry is reused and leads to us sending
traffic to the proxy when we shouldn't.

Thus, we need to flush the CT entries at the end of the first test run,
with --flush-ct.

Fixes: de192decf1cf9 ("ci-ipsec-e2e: Add IPsec key rotation test")
Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>