controller: CreateController method for creation without triggering
[ upstream commit 96ea098690e358728e461d5666a847a83f5861d6 ]

The UpdateController method performs either a creation, or updating
of a controller. If an update is performed the controller is immediately
triggered. In some cases we want to create the controller once without
triggering or modifying it if it already has been created.

Add a CreateController method for those use cases.

Signed-off-by: Jussi Maki <jussi@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>