docs: Add Makefile and documentation for "fast" development targets
[ upstream commit 5f88bbffeb50a3de415549a3a9d50e34453ccd34 ]

This commit introduces documentation for "fast" development targets in
Cilium. The "fast" targets allow for quicker development workflows by
building Cilium locally and deploying it on an existing kind cluster.
This approach eliminates the need to create and deploy Docker images
each time a source code change is made.

Additionally, the documentation has been updated to include these new
Makefile targets and provides a quick start guide with essential
information for working with Cilium.

When a cluster is already running with a Cilium container, we can
simply volume mount the binaries into the container and restart the
pod to pick up the changes. A similar approach can be used for C code,
with the caveat that the bpf/lib directory needs write permissions by
Cilium to generate the C Header features file, making it impossible
to use the volume as read-only. Since all nodes share the same kernel
features, there won't be any conflicts when multiple nodes write
to the same volume mount.

Signed-off-by: André Martins <andre@cilium.io>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>