ipam: report IP owner of non-default pool IPs in multi-pool IPAM
[ upstream commit afd8e4eef3b64411156a82d8c10187f70b2c00bf ]

Before this change, only IPs from the default pool are reported in
`cilium status --verbose` in the "Allocated addresses" section when
using multi-pool IPAM mode. With this change, all IPs from non-default
pools are reported as well by making use of the fact that the pool name
is prepended to the IP in `(*multiPoolManager).dump`.

Example:

```
% ks exec -it cilium-n8xw7 -- cilium status --verbose
[...]
IPAM:                   IPv4: 2 IPAM pool(s) available,
Allocated addresses:
  10.10.0.46 (router)
  10.10.0.48 (kube-system/coredns-5d78c9869d-d4nts [restored])
  10.10.0.61 (kube-system/hubble-relay-597b7bdff8-pwl69 [restored])
  10.10.0.62 (health)
  jupiter/192.168.0.15 (default/nginx-jupiter)
[...]
```

Note that IPs from non-default pools (in this case the `jupiter` pool)
are reported with the pool name prefixed.

Signed-off-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>