ci/multi-network: Fix pod CIDR range for secondary network
This commit fixes an issue where pod IPs for pods on the secondary
network were allocated from the `192.168.0.0/20` range, which overlaps
with the node IP range (`192.168.0.0/24`) set up by our `kind.sh` script
[1]. The two must not overlap, as otherwise pods and nodes might attempt
to use the same IP, which is fatal in direct routing mode.

Instead, we move the pod CIDR range to `192.168.16.0/20`, which does not
overlap with `192.168.0.0/24` anymore.

[1]: https://github.com/isovalent/cilium/blob/bed2b12f0ac322a7afab361933cfa3413dccae5f/contrib/scripts/kind.sh#L53

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>