bpf: egressgw-ha: select new gateway for CT_REOPENED
The EgressGW HA feature supports selecting multiple gateways, and we have
a per-connection BPF map that tracks the currently selected gateway (so
that all traffic for a given connection flows via the same gateway).
This map is reconciled by the agent, to delete entries for eg. stale
endpoints or unhealthy gateways.

As egress_gw_ha_request_needs_redirect() bails out early for CT_REPLY and
CT_RELATED results, we currently inspect this map for CT_ESTABLISHED and
CT_REOPENED results. For CT_NEW we just select a fresh gateway (even though
the map might still contain an earlier selection).

The CT lookup for a packet returns CT_REOPENED
- it is TCP with SYN set (to reach the ACTION_CREATE case), and
- it matches a previous outbound connection, and
- the connection has seen a FIN / RST packet

This seems like a good moment to rebalance the connection, and try to
select a different gateway. Doing so helps to drain gateways that are no
longer selected by policy, but still receive traffic for active
connections. It also provides more graceful failover in case that the
end-to-end network path via the currently selected gateway is disrupted
(without the gateway being detected as unhealthy).

Note that this won't help for "simple" SYN retries. Here the lookup
result is still CT_ESTABLISHED (as we haven't seen any FIN / RST).

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>