helm: add tolerations support to clustermesh certgen
[ upstream commit 5e3202619f4ab91e3eb5fe753667a3c3c03b1ebe ]

A certgen job can be used to generate the TLS certificates required
both for hubble and clustermesh. Yet, the clustermesh one lacks the
possibility of configuring additional tolerations, which are instead
configurable in the hubble case. This is possibly useful for instance
to allow running the job on control-plane nodes. Hence, let's fix this
divergence.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>