CI images: Define a variable for the floating tags
[ upstream commit 5b55b65a8dc0a56977699e80f37bb190f9c958e7 ]

Define a variable for CI image floating tags instead of hardcoding it.

- For the main branch, the floating tag is "latest".
- For stable branches, the floating tag is the same as the branch name
  (e.g. "v1.14").

This commit was initially introduced in #28008, but it got reverted
because it tried to push floating tags for feature branches and failed.
Now that #28044 is merged, this workflow no longer pushes floating tags
from feature branches. Therefore it is safe to use github.ref_name as a
floating tag.

Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>