ci: update k8s versions support for v1.14-ce
We offer extended support for newer K8s versions in Enterprise compared
to OSS (which sticks to whatever K8s versions were available at the time
of release).

At the moment, 1.25 - 1.27 are available on all the cloud providers, so
we'll just test against that to ensure compatibility (with 1.27 as the
default testing for any backport PRs).

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>