enterprise/cilium-cni: change multi-network label format
Currently, the Cilium CNI plugin attaches a label of the format

    cni:com.isovalent.v1alpha1.network.attachment/<network name>=<pod interface name>

to Cilium endpoints representing a secondary network attachment.
However, including the pod interface name is an implementation detail
and makes it harder for users to match the label in policy. Thus, change
the label to the format

    cni:com.isovalent.v1alpha1.network.attachment=<network name>

As a consequence, if a pod is attached to the same network multiple
times, the respective endpoints will now get assigned the same security
identiy which makes policy selection for networks more straight forward.

Suggested-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>