atlantis: Use FQDN proxy plugin from commit f6b36454e9
To adapt to changes in the agent introduced in upstream commits, the
fqdn-proxy plugin has been updated accordingly.

This commit updates the plugin version used to patch the agent with the
one that is compatible with the latest version of the agent code.

Relates: https://github.com/isovalent/fqdn-proxy/pull/813

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>