cilium: Print full labelset for all reserved identities
[ upstream commit dba3ee1eb9991e70471e37f795a3991662298a78 ]

In particular the host identity can have multiple labels that are not
hardcoded directly from the number, so we should just always fetch the
full labelset from the agent to ensure that the CLI prints what the
agent knows. No special cases for reserved identities.

The main subtlety here is that the GetIdentityIDParams() input is always
the numeric identity as a string, and never the stringified name of that
identity. For instance if client-side the CLI thinks that it can map
identity "1" to "host" then it will still make the query with the
parameter "1" (since that's how the identity get API works). If the
query fails, then it'll attempt to fall back to printing a stringified
name like "host" if available client-side, or otherwise just print the
number.

Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>