cilium: Fix variable naming for labels
[ upstream commit fafd535bbb623805f4144a74a61db10fafad2068 ]

A single IP can only have exactly one identity. Each identity can have a
set of labels associated with it. This code was confusingly using the
variable name 'identities' for individual labels within a single
identity. Rename it for clarity.

(Strictly speaking the labels may not actually be labels but may be the
numeric identity if the user decided not to pretty-print the labels, but
I think this variable naming is still more clear than what preceded it).

Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>