helm: don't expose kvstoremesh metrics if not enabled
[ upstream commit e1b69e75bffdcd2ae043766216091c2abfaeeeb2 ]

Currently, the kvstoremesh metrics entry inside the dedicated
Service and ServiceMonitor is always inserted if the corresponding
setting is enabled (true by default), regardless of whether
kvstoremesh is actually enabled or not. Yet, this would create
problems when Prometheus attempts to scrape it, as the target
port would be closed. Hence, let's update the helm chart so that
these entries are inserted only if kvstoremesh is enabled.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>