egressgw: Switch from net to netip
[ upstream commit 6ab32dfd35979e52d50a27b5d880247772b3bf4d ]

[ backporter's note: the the FromAddr() method in pkg/types/ipv4.go has been
reworked to replace the clear method, as that's not yet available in go 1.20 ]

Modernize this section of the codebase by switching over to newer
addressing types provided by recent Go standard libraries.

The most interesting aspect here is the use of netip.Addr{} as the
"invalid ip" value, which is distinct from netip.IPv4Unspecified, ie
"0.0.0.0". Both of these values are used at different times in egress
gateway code, and we must be careful not to conflate them.
addr.IsValid() will determine whether the IP is an actual IP, or if
false, the "invalid ip" uninitialized address.

Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Joe Stringer <joe@isovalent.com>

clear