ipcache: fix releasing node CIDRs after restoration
[ upstream commit 9016d93a0e35f50e82026e9499d76e156fd7130b ]

[ backporter's notes: removed reference to WorldV4 / V6 labels; they're
  not present in v1.14 ]

This fixes a bug where node CIDR identities were not cleaned up after
restoration. Specifically, the set of labels used for allocation vs. the
set of labels released may not be the same.

This is made more awkward by the fact that we don't actually *know* the
set of labels that were allocated -- only the CIDRs. This means we have
to go and recreate them. So, we guess. However, now that CIDR identities
can be either global *or* remote-node, we need to update our guess.

So, try and release CIDR identiites as usual, but if the node-cidr
feature is enabled, and the first guess is incorrect, also try in the
node-cidr style.

Fixes: 02406f2c
Fixes: 89350f75

Signed-off-by: Casey Callendrello <cdc@isovalent.com>