pkg/endpoint: run the metadata resolver after registering the endpoint
[ upstream commit 07d3a21635f491fb07d0f5b5ef0d146047a88c6d ]

We need to refetch the pod labels again because we have just added
the endpoint into the endpoint manager. If we have received any pod
events, more specifically any events that modified the pod labels,
between the time the pod was created and the time it was added
into the endpoint manager, the pod event would not have been processed
since the pod event handler would not find the endpoint for that pod
in the endpoint manager. Thus, we will fetch the labels again
and update the endpoint with these labels.

Signed-off-by: André Martins <andre@cilium.io>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>