policy: Policy Map Overflow Lockdown
This logic has an option (`enable-policy-overflow-lockdown-endpoint`),
"true" by default, so that it can be disabled.

When a policy map needs more entries than is
allocated for the bpf map, the endpoint is locked
down. If the endpoint has deny policies that can still
fit into the policy map then they will be insert into
the policy map with all allows removed. If there are
no deny policies or the deny policies cannot fit into
the map then the policy cache is reset and the policy map
is replaced with a set of default deny entries.
The endpoint is put into full default-deny for every protocol.

If the lockdown calls fail, then the endpoint will be
regenerated, and another lockdown will be attempted (until
the underlying map pressure is relieved). If the lockdown
mode is still inserting deny policy entries further
adjustments to the map may still happen (in the form of
deny policies being added and/or removed).

The Map structure in the bpf package now has a
Pin method and the policymap now has a CreateUnpinned
method so that it can be initialized before it is
created.

Incremental deletion is now done first in cases where
incremental addition before deletion would overflow the
policy map.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>