[not mindful] endpoint: reconstruct multi-network endpoint labels in metadata resolver
A recent change the way endpoints labels are updated, see commit
f398a6a14c48 ("pkg/endpoint: run the metadata resolver after registering
the endpoint").

This change causes labels added by the CNI no longer being added to the
endpoint labels. This breaks the multi-network feature because it relies
on CNI-added labels for the endpoints representing secondary network
attachments.

Work around this in a non-mindful manner for now to unblock the next
1.14 CEE release. The proper fix will involve some more work in Cilium
OSS. Once that fix is merged and backported, this workaround can be
dropped again.

Updates isovalent/cilium#2152

Co-authored-by: André Martins <andre@cilium.io>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>