pkg: Add Bitwise LPM Trie Library
This bitwise lpm trie is a non-thread-safe binary
trie that indexes arbitrarily long bit-based keys
with associated prefixes indexed from most
significant bit to least significant bit using
the longest prefix match algorithm.

Documenting the behavior of the datastructure is
localized around the method calls in the trie.go
file.

The tests specifically test boundary cases for the
various methods.

Updating CODEOWNERS to put sig-policy and ipcache
in charge of this library.

Fixes: #29519

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>